working arch optinaisation
